+++ oidc-client/internal/pkce/pkce.go
@@ -1,4 +1,4 @@
-package internal
+package pkce
 
 import (
 	"crypto/rand"
