+++ oidc-client/internal/pkce/pkce_test.go
@@ -1,4 +1,4 @@
-package internal
+package pkce
 
 import "testing"
